requirements.txt
Brotli==1.0.7
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-autoslug-iplweb==1.9.4
django-dash==0.5.6
django-heroku==0.3.1
django-nine==0.2.2
django-tinymce==2.8.0
djangorestframework==3.10.3
easy-thumbnails==2.6
feedparser==5.2.1
freeze==1.0.10
gunicorn==19.9.0
idna==2.8
lxml==4.4.1
numpy==1.17.3
ordereddict==1.1
pandas==0.25.2
pandas-datareader==0.8.1
pif==0.8.2
Pillow==6.2.0
psycopg2==2.8.4
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
transliterate==1.10.2
Unidecode==1.1.1
urllib3==1.25.6
vishap==0.1.5
whitenoise==4.1.4
